+++ packages/cactus-plugin-ledger-connector-corda/src/main-server/kotlin/gen/kotlin-spring/src/main/kotlin/org/hyperledger/cactus/plugin/ledger/connector/corda/server/model/DeployContractJarsV1Request.kt
@@ -22,7 +22,7 @@ data class DeployContractJarsV1Request(
 
     @get:NotNull  
     @field:Valid
-    @get:Size(min=1,max=1024)
+    @get:Size(min=0,max=1024)
     @field:JsonProperty("cordappDeploymentConfigs") val cordappDeploymentConfigs: kotlin.collections.List<CordappDeploymentConfig>,
 
     @get:NotNull
